[NEW] relocate spotify functions
